requirements.txt
# aws-cdk-lib==2.58.1
aws-cdk-lib==2.66.0
constructs>=10.0.0,<11.0.0
aws-cdk.lambda-layer-kubectl-v24
python-benedict==0.22.4